requirements.txt
Django==4.1.3
djangorestframework==3.14.0
uwsgi==2.0.20
drf-yasg==1.21.4
django-split-settings==1.1.0
django-filter==22.1
mysqlclient==2.1.1
djangorestframework-simplejwt[crypto]==5.2.2